requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
ipykernel
tensorflow
matplotlib
joblib
scikit-learn
comet-ml
dvc